+++ src/Products/Domain/ValueObjects/Quality.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\VillaPeruana\Products\Domain\ValueObjects;
+namespace App\Products\Domain\ValueObjects;
 
 final class Quality
 {
